requirements.txt
numpy==2.2.6
pandas==2.2.3
scipy==1.15.3
matplotlib==3.10.3
librosa==0.11.0
Pillow==11.2.1
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
